extract etag generator
