+++ backend/src/app.ts
@@ -9,11 +9,8 @@ import {
   rootRateLimiter,
 } from "./middlewares/rate-limit";
 import { compatibilityCheckMiddleware } from "./middlewares/compatibilityCheck";
-import {
-  COMPATIBILITY_CHECK,
-  COMPATIBILITY_CHECK_HEADER,
-} from "@monkeytype/contracts";
-import { default as etag } from "etag";
+import { COMPATIBILITY_CHECK_HEADER } from "@monkeytype/contracts";
+import { createETagGenerator } from "./utils/etag";
 
 const etagFn = createETagGenerator({ weak: true });
 
@@ -40,27 +37,6 @@ function buildApp(): express.Application {
   app.use(errorHandlingMiddleware);
 
   return app;
-}
-
-/**
- * create etag generator, based on the express implementation https://github.com/expressjs/express/blob/9f4dbe3a1332cd883069ba9b73a9eed99234cfc7/lib/utils.js#L247
- * Adds the api COMPATIBILITY_CHECK version in front of the etag.
- * @param options
- * @returns
- */
-function createETagGenerator(options: {
-  weak: boolean;
-}): (body: Buffer | string, encoding: BufferEncoding | undefined) => string {
-  return function generateETag(body, encoding) {
-    const buf = !Buffer.isBuffer(body) ? Buffer.from(body, encoding) : body;
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call
-    const generatedTag: string = etag(buf, options);
-    if (generatedTag.startsWith("W/")) {
-      return `W/"V${COMPATIBILITY_CHECK}-${generatedTag.slice(3)}`;
-    }
-    return `"V${COMPATIBILITY_CHECK}-${generatedTag.slice(1)}`;
-  };
 }
 
 export default buildApp();
